graphics/PlatformProbe: Factor out a `select_buffer_allocating_renderer`
