+++ src/server/graphics/default_configuration.cpp
@@ -359,75 +359,12 @@ mir::DefaultServerConfiguration::the_buffer_allocator()
     return buffer_allocator(
         [&]() -> std::shared_ptr<graphics::GraphicBufferAllocator>
         {
-            /* Theory:
-             * We iterate over each DisplaySink and each RenderingPlatform, and
-             * collate:
-             * *) If the RenderingPlatform is *best* at driving that sink, and
-             * *) If the RenderingPlatform is *unable* to drive that sink
-             *
-             * We then filter to just the RenderingPlatform(s) that are able to
-             * drive the most DisplaySink.
-             *
-             * Of that list, we pick the RenderingPlatform that has the highest
-             * number of “best” suitabilities.
-             */
-            struct PlatformData
-            {
-                std::shared_ptr<mg::RenderingPlatform> platform;
-                std::shared_ptr<mg::GLRenderingProvider> provider;
-                int best_count;
-                int unsupported_count;
-            };
-            std::vector<PlatformData> providers;
-            providers.reserve(the_rendering_platforms().size());
-            for (auto const& platform : the_rendering_platforms())
-            {
-                providers.push_back({
-                        platform,
-                        platform->acquire_provider<mg::GLRenderingProvider>(platform),
-                        0,
-                        0
-                });
-            }
-
-            the_display()->for_each_display_sync_group(
-                [&](auto& sync_group)
-                {
-                    sync_group.for_each_display_sink(
-                        [&](auto& sink)
-                        {
-                            for (auto &[platform, provider, best_count, unsupported_count] : providers)
-                            {
-                                auto suitability = provider->suitability_for_display(sink);
-                                if (suitability >= mg::probe::best)
-                                {
-                                    best_count++;
-                                }
-                                else if (suitability == mg::probe::unsupported)
-                                {
-                                    unsupported_count++;
-                                }
-                            }
-                        });
-                });
-
-            // providers.size() is guaranteed to be ≥ 1, as we're guaranteed to have at least one rendering platform
-
-            auto const least_unsupported_count = std::min_element(
-                providers.begin(), providers.end(),
-                [](auto const& a, auto const& b) { return a.unsupported_count < b.unsupported_count; })->unsupported_count;
-
-            auto const last_good_provider = std::remove_if(
-                providers.begin(), providers.end(),
-                [least_unsupported_count](auto const& provider) { return provider.unsupported_count > least_unsupported_count; });
-
-            // Now, find the best platform out of the ones that support the most Sinks
-            auto const best_provider = std::max_element(
-                providers.begin(), last_good_provider,
-                [](auto const& a, auto const&b) { return a.best_count < b.best_count; });
-
+            auto rendering_platforms = the_rendering_platforms();
+            auto best_provider = graphics::select_buffer_allocating_renderer(
+                *the_display(),
+                rendering_platforms);
             // TODO: More than one BufferAllocator
-            return best_provider->platform->create_buffer_allocator(*the_display());
+            return best_provider->create_buffer_allocator(*the_display());
         });
 }
